requirements.txt
streamlit==1.28.2
firebase-admin==6.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
plotly==5.17.0
pandas==2.1.3
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
pyrebase4==4.7.1
streamlit-javascript==0.1.5